Move exclusions to separate file
